fixing styling for church-creation page
