+++ app/church-creation/page.tsx
@@ -1,9 +1,11 @@
-import ChurchCreationForm from './components/ChurchCreationForm';
+import ChurchCreationForm from '../components/forms/ChurchCreationForm';
 
 export default function ChurchCreationPage() {
     return (
-        <div className="container mx-auto">
-            <ChurchCreationForm />
-        </div>
+        <section className="min-h-screen flex items-center justify-center bg-gradient-to-t from-blue-300 to-blue-600 pt-40">
+            <div className="w-full max-w-md">
+                <ChurchCreationForm />
+            </div>
+        </section>
     );
 }
